pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ishihara-dot-plate-generator"
version = "0.1.0"
description = "A modular Python app for generating Ishihara-style color vision test plates."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pillow",
    "customtkinter",
    "pytest",
    "pytest-cov",
    "numpy"
]

[tool.setuptools.packages.find]
where = ["src"]
